+++ tests/feature/test_multiline.py
@@ -84,10 +84,10 @@ def test_multiline(testdir, feature_text, expected_text):
 
             @scenario("multiline.feature", "Multiline step using sub indentation")
             def test_multiline(request):
-                assert request.getfixturevalue("i_have_text") == expected_text
+                assert request.getfixturevalue("text") == expected_text
 
 
-            @given(parsers.parse("I have a step with:\\n{{text}}"))
+            @given(parsers.parse("I have a step with:\\n{{text}}"), target_fixture="i_have_text")
             def i_have_text(text):
                 return text